987 fixed deselecting graph nodes
