+++ app/javascript/controllers/thematic_map_controller.js
@@ -1,19 +1,9 @@
-import { Controller } from "@hotwired/stimulus"
+import GraphController from "controllers/graph_controller";
 import * as d3 from "d3"
 
-export default class extends Controller {
+export default class extends GraphController {
 
   static targets = [
-    "map", 
-    "graph", 
-    "filterForm", 
-    "toggleLabelsButton", 
-    "dialog", 
-    "dialogContent", 
-    "dialogTitle",
-    "dialogTitleColor", 
-    "dialogContent",
-    "closeDialog",
     "stakeholders", 
     "initiatives",
     "selectedStakeholdersForPrint", 
@@ -23,8 +13,7 @@ export default class extends Controller {
   ]
 
   connect() {
-    window.addEventListener('beforeprint', this.beforePrint.bind(this));
-    window.addEventListener('afterprint', this.afterPrint.bind(this));
+    super.connect()
 
     const data = this.getData()
     this.drawGraph(data)
@@ -41,23 +30,6 @@ export default class extends Controller {
     this.updateLabelsVisibility()
   }
 
-  disconnect() {
-    window.removeEventListener('beforeprint', this.beforePrint.bind(this));
-    window.removeEventListener('afterprint', this.afterPrint.bind(this));
-  }
-
-  afterPrint() {
-    this.mapTarget.removeAttribute('height');
-  }
-
-  beforePrint() {
-    const svg = this.mapTarget.querySelector('svg');
-    const bounds = svg.getBBox();
-
-    this.mapTarget.setAttribute('height', bounds.height + 50);
-    svg.setAttribute('height', bounds.height + 50);
-  }
-
   drawGraph(data) {
     const width = this.mapTarget.clientWidth
     const height = this.mapTarget.clientHeight
@@ -72,11 +44,11 @@ export default class extends Controller {
     let dialogTitleTarget = this.dialogTitleTarget
     let dialogTitleColorTarget = this.dialogTitleColorTarget
     let dialogContentTarget = this.dialogContentTarget
+    let stakeholdersTarget = this.stakeholdersTarget
+    let initiativesTarget = this.initiativesTarget
 
     let querySelectorIncludesText = this.querySelectorIncludesText.bind(this)
-    // let updateStakeholderTypes = this.updateStakeholderTypes.bind(this)
-
-    // let stakeholderTypesTarget = this.stakeholderTypesTarget
+    let updateNodeColors = this.updateNodeColors.bind(this)
 
     var linkForce = d3
       .forceLink()
@@ -113,10 +85,12 @@ export default class extends Controller {
       .attr("height", height)
       .on('click', function(event) {
         if (event.target.tagName === 'svg') {
-          const selectedStakeholders = Array.from(this.stakeholdersTarget.selectedOptions).map(({ value }) => value)
-          const selectedInitiatives = Array.from(this.initiativesTarget.selectedOptions).map(({ value }) => value)
+          linkElements.attr('class', function (link) { 'links stroke-zinc-400 dark:stroke-zinc-400' })
 
-          this.updateNodeColors(selectedStakeholders, selectedInitiatives)
+          const selectedStakeholders = Array.from(stakeholdersTarget.selectedOptions).map(({ value }) => value)
+          const selectedInitiatives = Array.from(initiativesTarget.selectedOptions).map(({ value }) => value)
+
+          updateNodeColors(selectedStakeholders, selectedInitiatives)
         }
       });
 
@@ -277,99 +251,16 @@ export default class extends Controller {
     return { nodes: nodesData, links: linksData };
   }
 
-  querySelectorIncludesText(selector, text) {
-    return Array.from(document.querySelectorAll(selector))
-      .find(el => el.textContent.includes(text));
-  }
-
-
-  getNeighbors(links, node) {
-    return links.reduce(function (neighbors, link) {
-        if (link.target.id === node.id) {
-          neighbors.push(link.source.id)
-        } else if (link.source.id === node.id) {
-          neighbors.push(link.target.id)
-        }
-        return neighbors
-      },
-      [node.id]
-    )
-  }
-
-  getLinkClass(link, node) {
-    if (link.target.id === node.id || link.source.id === node.id) {
-      return 'links stroke-green-300 dark:stroke-green-300'
-    } else {
-      return 'links stroke-zinc-400 dark:stroke-zinc-400'
-    }
-  }
-
-  getNodeColor(node, neighbors) {
-    if (Array.isArray(neighbors) && neighbors.includes(node.id)) {
-      return '#49C472'
-    } else  {
-      return node.color
-    }
-  }
-
   getNodeSize(node) {
     var nodeSize = node.size
 
     return nodeSize;
   }
 
-  getTextClass(node, neighbors) {
-    if (Array.isArray(neighbors) && neighbors.includes(node.id)) {
-      return 'texts stroke-green-300 dark:stroke-green-300'
-    } else {
-      return 'texts stroke-zinc-400 dark:stroke-zinc-400'
-    }
-  }
-
   calcForceStrength(nodes, links) { return -40 }
 
   calcLinkStrength(links) {
     return 0.38
-  }
-
-  toggleLabels(event) {
-    if (typeof event.stopPropagation === 'function') {
-      event.stopPropagation()
-    }
-
-    const url = new URL(window.location)
-    const showLabels = url.searchParams.get('show_labels') == 'true'
-
-    if (showLabels) {
-      url.searchParams.delete('show_labels')
-    } else {
-      url.searchParams.append('show_labels', 'true')
-    }
-
-    window.history.replaceState({}, '', url)
-
-    this.updateLabelsVisibility()
-  }
-
-  updateLabelsVisibility() {
-    const textElements = this.mapTarget.querySelectorAll('.texts text')
-
-    const url = new URL(window.location)
-    const showLabels = url.searchParams.get('show_labels') == 'true'
-
-    if (showLabels) {
-      this.toggleLabelsButtonTarget.innerHTML = 'Hide names'
-
-      textElements.forEach(element => {
-        element.setAttribute('visibility', 'visible')
-      })
-    } else {
-      this.toggleLabelsButtonTarget.innerHTML = 'Show names'
-
-      textElements.forEach(element => {
-        element.setAttribute('visibility', 'hidden')
-      })
-    }
   }
 
   updateFilters(event) {
